Create Rock_Paper_Scissors.java
Simplest Rock Paper Scissors with the computer. Although not the best, it works.
So the thing is that the program is a bit lazy so it will only give out its Rock, Paper or Scissors only after you have selected your choice. But no problem cause it can not see your answer. Its just not punctual to give its choice on time.